♻️ refactor(SCRUM-242): PortfolioSideBar 분리 & 타입 네이밍 컨벤션에 따라 수정
